use float32 and complex64 instead of float64 and complex128

+++ npbench/benchmarks/azimint_hist/azimint_hist_numba_n.py
@@ -40,7 +40,7 @@ import numba as nb
 
 @nb.jit(nopython=True, parallel=False, fastmath=True)
 def get_bin_edges(a, bins):
-    bin_edges = np.zeros((bins + 1, ), dtype=np.float64)
+    bin_edges = np.zeros((bins + 1, ), dtype=np.float32)
     a_min = a.min()
     a_max = a.max()
     delta = (a_max - a_min) / bins
