+++ npbench/benchmarks/cavity_flow/cavity_flow.py
@@ -4,9 +4,9 @@ import numpy as np
 
 
 def initialize(ny, nx):
-    u = np.zeros((ny, nx), dtype=np.float64)
-    v = np.zeros((ny, nx), dtype=np.float64)
-    p = np.zeros((ny, nx), dtype=np.float64)
+    u = np.zeros((ny, nx), dtype=np.float32)
+    v = np.zeros((ny, nx), dtype=np.float32)
+    p = np.zeros((ny, nx), dtype=np.float32)
     dx = 2 / (nx - 1)
     dy = 2 / (ny - 1)
     dt = .1 / ((nx - 1) * (ny - 1))
