+++ npbench/benchmarks/contour_integral/contour_integral_numba_npr.py
@@ -6,12 +6,12 @@ import numba as nb
 
 @nb.jit(nopython=True, parallel=True, fastmath=True)
 def contour_integral(NR, NM, slab_per_bc, Ham, int_pts, Y):
-    P0 = np.zeros((NR, NM), dtype=np.complex128)
-    P1 = np.zeros((NR, NM), dtype=np.complex128)
+    P0 = np.zeros((NR, NM), dtype=np.complex64)
+    P1 = np.zeros((NR, NM), dtype=np.complex64)
     # for z in int_pts:
     for i in nb.prange(len(int_pts)):
         z = int_pts[i]
-        Tz = np.zeros((NR, NR), dtype=np.complex128)
+        Tz = np.zeros((NR, NR), dtype=np.complex64)
         for n in nb.prange(slab_per_bc + 1):
             zz = np.power(z, slab_per_bc / 2 - n)
             Tz += zz * Ham[n]
