+++ npbench/benchmarks/contour_integral/contour_integral_pythran.py
@@ -3,13 +3,13 @@
 import numpy as np
 
 
-# pythran export contour_integral(int, int, int, complex128[:,:,:],
-#                                complex128[:], complex[:,:])
+# pythran export contour_integral(int, int, int, complex64[:,:,:],
+#                                complex64[:], complex[:,:])
 def contour_integral(NR, NM, slab_per_bc, Ham, int_pts, Y):
-    P0 = np.zeros((NR, NM), dtype=np.complex128)
-    P1 = np.zeros((NR, NM), dtype=np.complex128)
+    P0 = np.zeros((NR, NM), dtype=np.complex64)
+    P1 = np.zeros((NR, NM), dtype=np.complex64)
     for z in int_pts:
-        Tz = np.zeros((NR, NR), dtype=np.complex128)
+        Tz = np.zeros((NR, NR), dtype=np.complex64)
         for n in range(slab_per_bc + 1):
             zz = np.power(z, slab_per_bc / 2 - n)
             Tz += zz * Ham[n]
