+++ npbench/benchmarks/polybench/k3mm/k3mm_dace.py
@@ -6,7 +6,7 @@ NI, NJ, NK, NL, NM = (dc.symbol(s, dtype=dc.int64)
 
 
 @dc.program
-def kernel(A: dc.float64[NI, NK], B: dc.float64[NK, NJ], C: dc.float64[NJ, NM],
-           D: dc.float64[NM, NL]):
+def kernel(A: dc.float32[NI, NK], B: dc.float32[NK, NJ], C: dc.float32[NJ, NM],
+           D: dc.float32[NM, NL]):
 
     return A @ B @ C @ D
