+++ npbench/benchmarks/polybench/mvt/mvt_dace.py
@@ -5,8 +5,8 @@ N = dc.symbol('N', dtype=dc.int64)
 
 
 @dc.program
-def kernel(x1: dc.float64[N], x2: dc.float64[N], y_1: dc.float64[N],
-           y_2: dc.float64[N], A: dc.float64[N, N]):
+def kernel(x1: dc.float32[N], x2: dc.float32[N], y_1: dc.float32[N],
+           y_2: dc.float32[N], A: dc.float32[N, N]):
 
     x1 += A @ y_1
     x2 += y_2 @ A
